requirements.txt
Pillow==10.4.0
pytesseract==0.3.13
pandas==2.2.3
openai>=1.0.0
easyocr>=1.7.0
torch>=2.0.0
torchvision>=0.15.0
openpyxl>=3.1.0

